(Navbar) Navbar Desktop - implement navbar desktop
