+++ frontend/src/components/layout/NavbarDesktop/NavbarDesktopOld.module.css
@@ -1,0 +1,198 @@
+.navbarRoot {
+    --navbar-bg: #222755;
+    --navbar-text: #ffffff;
+    --primary-blue: #0173BE;
+    --card-bg: #E7F2FF;
+    --card-text: #222755;
+    --hover-fill: rgba(143, 210, 254, 0.25);
+    --active-bg: rgba(1, 115, 190, 0.40);
+
+    --navbar-height: 5.9375rem;
+    --navbar-padding: 2rem;
+    --pill-height: 3.625rem;
+    --pill-padding: 1.875rem;
+    --card-width: 18.9375rem;
+    --card-height: 7.8125rem;
+
+    /* Effects */
+    --border-radius: 1rem;
+    --shadow-light: 0 0.75rem 1.5rem rgba(0, 0, 0, 0.12);
+    --shadow-heavy: 0 0.625rem 1.5625rem rgba(0, 0, 0, 0.25);
+    --transition: 150ms;
+}
+
+/* BASE NAVBAR*/
+.navbarRoot { 
+    position: fixed; 
+    z-index: 9999; 
+    overflow: visible; 
+    top: 0;
+    left: 0;
+    width: 100%;
+    padding-top: 3rem; 
+}
+
+.navbar {
+    background-color: var(--navbar-bg);
+    border-radius: var(--border-radius);
+    box-shadow: var(--shadow-heavy);
+    overflow: visible;
+    user-select: none;
+    max-width: 94.5rem;
+    width: 95%;
+    margin: 0 auto;
+}
+
+.navbarBar {
+    height: var(--navbar-height);
+    display: flex;
+    align-items: center;
+    color: var(--navbar-text);
+    overflow: visible;
+    padding-left: var(--navbar-padding);
+    padding-right: var(--navbar-padding);
+}
+
+/* LOGO */
+.navbarLogo { display: inline-flex; align-items: center; margin-right: 4rem; flex-shrink: 0; text-decoration: none; }
+.navbarLogoImg { height: 4.6875rem; width: 4.6875rem; display: block; }
+
+/* NAV MENU */
+.navbarCenter { flex: 1; display: flex; justify-content: center; }
+.navbarList { display: flex; align-items: center; gap: 2rem; list-style: none; margin: 0; }
+.navbarItem { position: relative; overflow: visible; }
+
+/* NAVIGATION PILLS */
+.navPill {
+    position: relative;
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    gap: 0.5rem;
+    padding: 0 var(--pill-padding);
+    height: var(--pill-height);
+    border-radius: var(--border-radius);
+    font-weight: 600;
+    font-size: 1.25rem;
+    line-height: 1.875rem;
+    transition: all var(--transition);
+    outline: none;
+    border: 1px solid transparent;
+    background: transparent;
+    color: var(--navbar-text);
+    white-space: nowrap;
+    cursor: pointer;
+    text-decoration: none;
+}
+
+/* Pill glow effect */
+.navPill::before {
+    content: "";
+    position: absolute;
+    inset: -0.375rem;
+    border-radius: 1.375rem;
+    background: var(--primary-blue);
+    filter: blur(0.25rem);
+    opacity: 0.4;
+    z-index: -1;
+    pointer-events: none;
+}
+
+.navPill:hover { color: var(--navbar-text); }
+.navPill:focus-visible { outline: 2px solid rgba(0, 0, 0, 0.6); outline-offset: 2px; }
+
+/* Active state */
+.isActive {
+    background: var(--active-bg);
+    color: var(--navbar-text);
+    border: 1px solid rgba(0, 0, 0, 0.4);
+    backdrop-filter: blur(0.375rem);
+    -webkit-backdrop-filter: blur(0.375rem);
+}
+
+.isActive::before { inset: -0.5rem; border-radius: 1.5rem; filter: blur(1rem); }
+
+/* CARET ICON */
+.navCaret { height: 1.0625rem; width: 1.0625rem; transition: transform var(--transition); }
+.rotated { transform: rotate(180deg); }
+
+/* CTA BUTTON */
+.eventsBtn {
+    width: 9.375rem;
+    height: var(--pill-height);
+    border-radius: var(--border-radius);
+    background: var(--navbar-text);
+    color: #0e1a4b;
+    font-size: 1.25rem;
+    line-height: 1.875rem;
+    font-weight: 600;
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    transition: background-color var(--transition);
+    margin-left: 2rem;
+    flex-shrink: 0;
+    text-decoration: none;
+}
+
+.joinBtn:hover { background: #d7e9f9; }
+
+/* DROPDOWN SYSTEM */
+.dropdownOverlay { position: fixed; inset: 0; z-index: 9998; pointer-events: none; }
+.dropdownPanelWrap { position: absolute; pointer-events: auto; }
+.dropdownPanel { border-radius: var(--border-radius); background: var(--card-bg); box-shadow: var(--shadow-light); padding: 2.8125rem; position: relative; }
+.dropdownPointer { position: absolute; top: -2.15625rem; transform: translateX(-50%); pointer-events: none; }
+.pointerSvg { display: block; }
+
+/* DROPDOWN CARDS */
+.menuGrid { display: grid; grid-template-columns: 1fr 1fr; gap: 1rem; margin: 0; padding: 0; }
+.menuGridItem { list-style: none; }
+
+.navCard {
+    position: relative;
+    display: flex;
+    align-items: center;
+    gap: 1.1875rem;
+    width: var(--card-width);
+    height: var(--card-height);
+    padding: 1.3125rem;
+    border-radius: var(--border-radius);
+    text-decoration: none;
+    outline: none;
+    background: var(--card-bg);
+    transition: all var(--transition);
+    border: 2px solid transparent;
+    background-clip: padding-box;
+    isolation: isolate;
+}
+
+/* Card hover effects */
+.navCard:hover,
+.navCard:focus-visible {
+    background:
+            linear-gradient(var(--card-bg), var(--card-bg)) padding-box,
+            linear-gradient(to right, var(--navbar-bg), var(--primary-blue)) border-box;
+    border: 2px solid transparent;
+}
+
+.hoverFill {
+    position: absolute;
+    inset: 0;
+    border-radius: var(--border-radius);
+    background: var(--hover-fill);
+    opacity: 0;
+    transition: opacity var(--transition);
+    pointer-events: none;
+    z-index: 0;
+}
+
+.navCard:hover .hoverFill,
+.navCard:focus-visible .hoverFill { opacity: 1; }
+
+/* Card content */
+.iconWrap,
+.label { position: relative; z-index: 1; }
+.iconWrap { display: inline-flex; align-items: center; justify-content: center; height: 3rem; width: 3rem; }
+.iconImg { height: 3rem; width: 3rem; display: block; }
+.label { font-weight: 600; font-size: 1.25rem; line-height: 1.875rem; color: var(--card-text); }
+.ringGradient { position: absolute; inset: 0; border-radius: var(--border-radius); z-index: 0; pointer-events: none; }
